backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1

# HTTP 请求
requests>=2.31.0

# 文档解析
PyPDF2>=3.0.0
python-docx>=1.1.0

# LLM 和 Embedding
openai>=1.0.0
anthropic>=0.18.0

# 向量数据库
lancedb>=0.10.0

# Elasticsearch (可选，用于 BM25)
elasticsearch>=8.0.0

# 系统监控
psutil>=5.9.0

# Pydantic v2
pydantic>=2.0.0
